requirements.txt
arrow==0.12.1
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2018.10.15
chardet==3.0.4
clr==1.0.3
cycler==0.10.0
DateTime==4.3
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.1
django-admin-interface==0.5.9
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-chartjs==1.3
django-colorfield==0.1.14
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==1.1.0
django-jinja2==0.1
django-jinja2loader==0.1.2
django-modelcluster==4.2
django-tables2==2.0.1
django-taggit==0.23.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
docopt==0.6.2
draftjs-exporter==2.1.3
EasyTkinter==1.1.0
enum34==1.1.6
env==0.1.0
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==0.9999999
idna==2.7
jdcal==1.4
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
maze==1.3
numpy==1.13.3
odfpy==1.3.6
olefile==0.44
opencv-python==3.4.1.15
openpyxl==2.5.8
pandas==0.21.0
Pillow==4.3.0
pipreqs==0.4.9
protobuf==3.4.0
pyparsing==2.2.0
PyPI-Browser==1.5
pystache==0.5.4
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.13
reportlab==3.4.0
requests==2.20.0
six==1.11.0
static==1.1.1
static-foundation==6.3.1.1
tablib==0.11.5
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.24
wagtail==2.3
Werkzeug==0.12.2
whitenoise==4.1
Willow==1.1
xlrd==1.1.0
xlwt==1.3.0
yarg==0.1.9
zope.interface==4.5.0
